chore: refactor for readibility and add comments.
